update lsp-types and change types (#303)
* Update lsp-types requirement from >=0.70, <0.84 to >=0.84, <0.85

* fix breaking changes

* update changelog

+++ codespan-lsp/src/lib.rs
@@ -27,8 +27,8 @@ fn location_to_position(
         Err(Error::InvalidCharBoundary { given })
     } else {
         let line_utf16 = line_str[..column].encode_utf16();
-        let character = line_utf16.count() as u64;
-        let line = line as u64;
+        let character = line_utf16.count() as u32;
+        let line = line as u32;
 
         Ok(LspPosition { line, character })
     }
@@ -77,7 +77,7 @@ where
     })
 }
 
-pub fn character_to_line_offset(line: &str, character: u64) -> Result<usize, Error> {
+pub fn character_to_line_offset(line: &str, character: u32) -> Result<usize, Error> {
     let line_len = line.len();
     let mut character_offset = 0;
 
@@ -90,7 +90,7 @@ pub fn character_to_line_offset(line: &str, character: u64) -> Result<usize, Err
             return Ok(line_len - chars_off - ch_off);
         }
 
-        character_offset += ch.len_utf16() as u64;
+        character_offset += ch.len_utf16() as u32;
     }
 
     // Handle positions after the last character on the line
